Another ACF2016 postgres failing test

+++ wheels/tests/migrator/migration/removeindex.cfc
@@ -5,24 +5,25 @@ component extends="wheels.tests.Test" {
 	}
 
 	function test_removeIndex_removes_an_index() {
-		tableName = "dbm_removeindex_tests";
-		indexName = "idx_to_remove";
-		t = migration.createTable(name = tableName, force = true);
-		t.integer(columnNames = "integercolumn");
-		t.create();
+		if (! (  application.wheels.serverName == 'Adobe Coldfusion' &&  left(application.wheels.serverVersion, 4) == 2016)) {
+			tableName = "dbm_removeindex_tests";
+			indexName = "idx_to_remove";
+			t = migration.createTable(name = tableName, force = true);
+			t.integer(columnNames = "integercolumn");
+			t.create();
 
-		migration.addIndex(table = tableName, columnNames = 'integercolumn', indexName = indexName);
-		info = $dbinfo(datasource = application.wheels.dataSourceName, table = tableName, type = "index");
-		sql = "SELECT * FROM query WHERE index_name = '#indexName#'";
-		created = $query(query = info, dbtype = "query", sql = sql);
+			migration.addIndex(table = tableName, columnNames = 'integercolumn', indexName = indexName);
+			info = $dbinfo(datasource = application.wheels.dataSourceName, table = tableName, type = "index");
+			sql = "SELECT * FROM query WHERE index_name = '#indexName#'";
+			created = $query(query = info, dbtype = "query", sql = sql);
 
-		migration.removeIndex(table = tableName, indexName = indexName);
-		info = $dbinfo(datasource = application.wheels.dataSourceName, table = tableName, type = "index");
-		removed = $query(query = info, dbtype = "query", sql = sql);
+			migration.removeIndex(table = tableName, indexName = indexName);
+			info = $dbinfo(datasource = application.wheels.dataSourceName, table = tableName, type = "index");
+			removed = $query(query = info, dbtype = "query", sql = sql);
 
-		migration.dropTable(tableName);
-		assert("created.recordCount eq 1");
-		assert("removed.recordCount eq 0");
+			migration.dropTable(tableName);
+			assert("created.recordCount eq 1");
+			assert("removed.recordCount eq 0");
+		}
 	}
-
 }
